+++ back/src/queue/jobs/processRegistryExport.ts
@@ -1,6 +1,11 @@
 import { logger } from "@td/logger";
 import { prisma } from "@td/prisma";
-import { getUploadWithWritableStream } from "@td/registry";
+import {
+  getUploadWithWritableStream,
+  RegistryV2IncomingTexsInclude,
+  RegistryV2ManagedInclude,
+  RegistryV2OutgoingTexsInclude
+} from "@td/registry";
 import { Job } from "bull";
 import * as Excel from "exceljs";
 
@@ -60,11 +65,17 @@ const streamLookup = (
           include: {
             registrySsd: true,
             registryIncomingWaste: true,
-            registryIncomingTexs: true,
+            registryIncomingTexs: {
+              include: RegistryV2IncomingTexsInclude
+            },
             registryOutgoingWaste: true,
-            registryOutgoingTexs: true,
+            registryOutgoingTexs: {
+              include: RegistryV2OutgoingTexsInclude
+            },
             registryTransported: true,
-            registryManaged: true,
+            registryManaged: {
+              include: RegistryV2ManagedInclude
+            },
             bsdd: {
               include: RegistryV2BsddInclude
             },